Tests/LibWeb/Text/expected/canvas/webgpu-get-instance.txt

gpu instanceof GPU: true
gpu is null: false
